fear(header): structure header with components
